Refactor EntityQuerySet.for_project_locale as get_entities_for_project_locale
